requirements.txt
alembic==1.10.3
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
tzdata==2023.3
Werkzeug==2.2.3
WTForms==3.0.1
